web/requirements-dev.txt
-r requirements.txt
selenium
pylint
autopep8
pylint-django
